Search for IP address in hostname.
Nearly all dynamic IPs have hostnames containing (part of) the IP address.

+++ fmt_ulong.c
@@ -1,9 +1,11 @@
 #include "fmt.h"
 
-unsigned int fmt_ulong(register char *s,register unsigned long u)
+size_t fmt_ulong(register char *s,register unsigned long u)
 {
-  register unsigned int len; register unsigned long q;
-  len = 1; q = u;
+  register size_t len;
+  register unsigned long q;
+  len = 1;
+  q = u;
   while (q > 9) { ++len; q /= 10; }
   if (s) {
     s += len;
